+++ perl/FSM/Scheduler/ISF.pm
@@ -9,15 +9,17 @@ no warnings 'experimental::signatures';
 use FSM::Debug;
 use FSM::Scheduler::ISF::LoweringIR;
 use FSM::Scheduler::ISF::Emitter::FSM;
+use FSM::Scheduler::ISF::Emitter::JSON;
 
 sub new($class, %args) {
     my $debug = $args{debug} // 0;
     fsm_trace_enter('Initialize ISF scheduler', 2);
 
     my $self = bless {
-        debug    => $debug,
-        ir       => FSM::Scheduler::ISF::LoweringIR->new(debug => $debug),
-        emitter  => FSM::Scheduler::ISF::Emitter::FSM->new,
+        debug       => $debug,
+        ir          => FSM::Scheduler::ISF::LoweringIR->new(debug => $debug),
+        fsm_emitter => FSM::Scheduler::ISF::Emitter::FSM->new,
+        json_emitter=> FSM::Scheduler::ISF::Emitter::JSON->new,
     }, $class;
 
     fsm_trace_exit('ISF scheduler initialized', 2);
@@ -28,10 +30,20 @@ sub lower($self, $actor) {
     fsm_trace_enter("Scheduler lower: $actor->{actor_name}", 2);
 
     my $ir  = $self->{ir}->build_module($actor);
-    my $fsm = $self->{emitter}->emit($ir);
+    my $fsm = $self->{fsm_emitter}->emit($ir);
 
     fsm_trace_exit("Scheduler lower completed", 2);
     return $fsm;
+}
+
+sub report($self, $actor) {
+    fsm_trace_enter("Scheduler report: $actor->{actor_name}", 2);
+
+    my $ir   = $self->{ir}->build_module($actor);
+    my $json = $self->{json_emitter}->emit($ir);
+
+    fsm_trace_exit("Scheduler report completed", 2);
+    return $json;
 }
 
 1;